Update Composer command causing issues with separate vendor directory
Remove the --classmap-authoritative argument from the template for new Vapor environments, as this creates issues when using the separate-vendor option.

Due to separate-vendor not updating the autoload classmaps for Composer on each deploy, any time that new classes are added, Composer is unable to find them using PSR-4 due to this classmap-authoritative argument. In the Composer docs for this argument (https://getcomposer.org/doc/articles/autoloader-optimization.md#optimization-level-2-a-authoritative-class-maps), it specifically states that "This option is very simple, it says that if something is not found in the classmap, then it does not exist and the autoloader should not attempt to look on the filesystem according to PSR-4 rules".