Check for envrionment-specific .env file
Keep parity with `vapor env:pull` which outputs to [`.env.{environment}`](https://docs.vapor.build/1.0/projects/environments.html#environment-variables). Ideally, there should be an option to pull the environment variables from the staging environment and use those for the build rather than relying on having an `.env` file. (I haven't seen that this is an option.)