Revert recent changes resulting in problems with mix-manifest.json
This is only triage to get Laravel Mix working again. Additional work
may be required to get Livewire's `manifest.json` working correctly.

This reverts #67 f2e34ab3d6cf4b4ee5e917c83b56aea6ae0d0682 which was an
attempt to fix #35 758214a02c1422b99997df8f608820add37d6c88.

The end result of #67 was mix-manifest.json being included in the asset
artifact and being removed from the code deployment artifact. This meant
Laravel Mix applications would fail as they expect to be able to read
the contents of `public/mix-manifest.json` at runtime.

This does not attempt to fix the original issue #35 was trying to
address. #35 would have included `public/manifest.json` in the code
deployment artifact (the desired behavior) but #67 would ensure both
`public/manifest.json` and `public/mix-manifest.json` would end up in
the asset artifact.

Unclear if `manifest.json` is needed by both client (so needs to be a
part of the asset artifact) and code (so also needs to be in the code
deployment artifact) so not trying to make that decision now.